Refactor SharedArrayBuffer buffered IO
