+++ src/buffered_io/sab.ts
@@ -1,0 +1,8 @@
+export namespace SAB {
+  // Indexes into Int32Arrays.
+  export const MAIN = 0;
+  export const WORKER = 1;
+  export const LENGTH = 2;
+  export const START = 3;
+  export const maxChars: number = 64; // Max number of characters that can be stored in the buffer.
+}
